fix: change InternalId type from int to int64 in TransactionFull struct because code fails on int32 systems

+++ transactions/transactions.go
@@ -188,7 +188,7 @@ type TransactionFull struct {
 	// Min: 1
 	InstallmentsCount *int `json:"installments_count,omitempty"`
 	// Internal unique ID of the transaction on the SumUp platform.
-	InternalId *int `json:"internal_id,omitempty"`
+	InternalId *int64 `json:"internal_id,omitempty"`
 	// Latitude value from the coordinates of the payment location (as received from the payment terminal reader).
 	// Min: 0
 	// Max: 90
